tests/requirements.txt
pytest==9.0.0
pytest-asyncio==1.3.0
pytest-homeassistant-custom-component==0.13.316
